Fix gRPC service test cleanup
