+++ mypy.ini
@@ -24,6 +24,14 @@ ignore_errors = True
 
 [mypy-ume.api]
 ignore_errors = True
+[mypy-ume.graph_routes]
+ignore_errors = True
+[mypy-ume.vector_routes]
+ignore_errors = True
+[mypy-ume.policy_routes]
+ignore_errors = True
+[mypy-ume.api_deps]
+ignore_errors = True
 
 [mypy-ume.config]
 ignore_errors = True
@@ -32,4 +40,12 @@ ignore_errors = True
 ignore_errors = True
 
 [mypy-ume.agent_orchestrator]
+ignore_errors = True
+[mypy-ume.tweet_bot]
+ignore_errors = True
+[mypy-ume.graph_schema]
+ignore_errors = True
+[mypy-ume.watchers.dev_log_watcher]
+ignore_errors = True
+[mypy-ume.tracing]
 ignore_errors = True
